REF: Deduplicate top-level package
Instead of having duplicate requirements and build sections, just
name the top-level package as the first output because the second
output is just a metapackage. This makes the recipe smaller, easier
to read, and actually includes the tests with importlib_resources.

Could also make the metapackage the top-level package which is
what is described in the docs as "implicit metapackage".

https://docs.conda.io/projects/conda-build/en/latest/resources/define-metadata.html#implicit-metapackages